Added debug helper
use echo var_chart($var)

whenever you need to see inside an object or array or whatever.

I'm loading the debug helper in the site controller constructor, so you have the function globally.

It echos to screen in a human readable format, and logs to the dev console as well.